Migrate channel-authenticity onto the shared tl_data seam
Deletes the skill's private wrapper copy: data access now goes through
skills/_shared/tl_data.py (stdin-passed queries, exit-code failure
classification), and the skill keeps only its own channel resolution in
scripts/channel_lookup.py. The three ES call sites read the CLI's flat
rows directly instead of a synthesized hits.hits shape, and queries move
off argv so big bodies can't hit argument-length limits.
